ai_registry/requirements.txt
asgiref==3.10.0
asttokens==3.0.0
black==25.9.0
boto3==1.40.68
botocore==1.40.68
click==8.3.0
decorator==5.2.1
Django==5.2.8
django-storages==1.14.6
django-taggit==6.1.0
executing==2.2.1
gunicorn==23.0.0
ipython==9.7.0
ipython_pygments_lexers==1.1.1
isort==7.0.0
jedi==0.19.2
jmespath==1.0.1
matplotlib-inline==0.2.1
mypy_extensions==1.1.0
packaging==25.0
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prompt_toolkit==3.0.52
psycopg==3.2.12
ptyprocess==0.7.0
pur==7.3.3
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytokens==0.3.0
s3transfer==0.14.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
